refactor: 도메인 정보 수정
