+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -1,9 +1,11 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
+import roomescape.domain.theme.Theme;
 import roomescape.domain.member.Member;
+import roomescape.domain.reservationtime.ReservationTime;
 
 public class Reservation {
