+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
+import roomescape.domain.reservation.Reservation;
 
 public interface ReservationRepository {
