+++ src/main/java/roomescape/domain/reservationtime/AvailableReservationTimeDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain.dto;
+package roomescape.domain.reservationtime;
 
 import java.time.LocalTime;
