+++ src/main/java/roomescape/domain/theme/ThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.theme;
 
 import java.time.LocalDate;
 import java.util.List;
