+++ src/main/java/roomescape/dto/request/ReservationTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservationtime.ReservationTime;
 
 public record ReservationTimeRequest(
         @NotBlank(message = "예약 시작 시간을 입력해주세요.")
