+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -2,7 +2,7 @@ package roomescape.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import roomescape.domain.Theme;
+import roomescape.domain.theme.Theme;
 
 public record ThemeRequest(
         @NotBlank(message = "테마 이름을 입력해주세요.")
