+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE theme
 (
     id          BIGINT       NOT NULL AUTO_INCREMENT,
-    name        VARCHAR(255) NOT NULL UNIQUE,
+    name        VARCHAR(30)  NOT NULL UNIQUE,
     description VARCHAR(255) NOT NULL,
     thumbnail   VARCHAR(255) NOT NULL,
     PRIMARY KEY (id)
@@ -9,8 +9,8 @@ CREATE TABLE theme
 
 CREATE TABLE reservation_time
 (
-    id       BIGINT       NOT NULL AUTO_INCREMENT,
-    start_at VARCHAR(255) NOT NULL,
+    id       BIGINT NOT NULL AUTO_INCREMENT,
+    start_at TIME   NOT NULL,
     PRIMARY KEY (id)
 );
 
@@ -19,18 +19,18 @@ CREATE TABLE member
     id       BIGINT       NOT NULL AUTO_INCREMENT,
     email    VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    name     VARCHAR(255) NOT NULL,
+    name     VARCHAR(30)  NOT NULL,
     role     ENUM ('ADMIN', 'USER') DEFAULT 'USER' NOT NULL,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE reservation
 (
-    id        BIGINT       NOT NULL AUTO_INCREMENT,
-    date      VARCHAR(255) NOT NULL,
-    member_id BIGINT       NOT NULL,
-    time_id   BIGINT       NOT NULL,
-    theme_id  BIGINT       NOT NULL,
+    id        BIGINT NOT NULL AUTO_INCREMENT,
+    date      DATE   NOT NULL,
+    member_id BIGINT NOT NULL,
+    time_id   BIGINT NOT NULL,
+    theme_id  BIGINT NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (member_id) REFERENCES member (id),
     FOREIGN KEY (time_id) REFERENCES reservation_time (id),
